+++ Devnot.Mentor.Api/Services/Interfaces/IUserService.cs
@@ -1,7 +1,7 @@
 ﻿using System.Threading.Tasks;
 using DevnotMentor.Api.Common.Response;
+using DevnotMentor.Api.CustomEntities.Auth;
 using DevnotMentor.Api.CustomEntities.Dto;
-using DevnotMentor.Api.CustomEntities.OAuth;
 using DevnotMentor.Api.Utilities.Security.Token;
 
 namespace DevnotMentor.Api.Services.Interfaces
